system-plugin: move data code to plugin/data directory
