+++ src/plugin/api/system/authentication/change.c
@@ -14,10 +14,10 @@
 #include "plugin/common.h"
 #include "libyang/tree_data.h"
 #include "plugin/api/system/authentication/store.h"
-#include "system/data/authentication/authorized_key.h"
-#include "system/data/authentication/authorized_key/list.h"
-#include "system/data/authentication/local_user.h"
-#include "system/data/authentication/local_user/list.h"
+#include "plugin/data/system/authentication/authorized_key.h"
+#include "plugin/data/system/authentication/authorized_key/list.h"
+#include "plugin/data/system/authentication/local_user.h"
+#include "plugin/data/system/authentication/local_user/list.h"
 #include "plugin/types.h"
 #include "umgmt/db.h"
 #include "umgmt/types.h"
