+++ src/plugin/api/system/authentication/check.c
@@ -13,8 +13,8 @@
 #include "check.h"
 #include "load.h"
 #include "plugin/common.h"
-#include "system/data/authentication/local_user/list.h"
-#include "system/data/authentication/authorized_key/list.h"
+#include "plugin/data/system/authentication/local_user/list.h"
+#include "plugin/data/system/authentication/authorized_key/list.h"
 
 #include <sysrepo.h>
 #include <utlist.h>
