+++ src/plugin/data/system/authentication/authorized_key.h
