+++ src/plugin/subscription/change.c
@@ -19,10 +19,10 @@
 #include "srpc/ly_tree.h"
 #include "sysrepo_types.h"
 #include "plugin/api/system/authentication/load.h"
-#include "system/data/authentication/local_user/list.h"
-#include "system/data/dns_resolver/search/list.h"
-#include "system/data/dns_resolver/server/list.h"
-#include "system/data/ntp/server/list.h"
+#include "plugin/data/system/authentication/local_user/list.h"
+#include "plugin/data/system/dns_resolver/search/list.h"
+#include "plugin/data/system/dns_resolver/server/list.h"
+#include "plugin/data/system/ntp/server/list.h"
 #include "plugin/types.h"
 #include "umgmt/db.h"
